penambahan activity & permission controller,  tambah role seeder, fix bug log controller

+++ app/Http/Controllers/Master/CompanyController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers\Master;
 
 use App\Http\Controllers\Controller;
 use App\Models\Company;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 use Exception;
 use Illuminate\Http\Request;
 use Yajra\DataTables\DataTables;
@@ -27,11 +29,16 @@ class CompanyController extends Controller
         $company = Company::orderByDesc('id')->get();
 
         return DataTables::of($company)->addIndexColumn()->addColumn('action', function ($data) {
+            $userauth = User::with('roles')->where('id', Auth::id())->first();
             $button = '';
-            $button .= ' <a href="' . route('company.edit', ['id' => $data->id]) . '" class="btn btn-sm btn-success action mr-1" data-id=' . $data->id . ' data-type="edit" data-toggle="tooltip" data-placement="bottom" title="Edit Data"><i
-            class="fas fa-pencil-alt"></i></a>';
-            $button .= ' <button  class="btn btn-sm btn-danger  action" data-id=' . $data->id . ' data-type="delete" data-route="' . route('company.delete', ['id' => $data->id]) . '" data-toggle="tooltip" data-placement="bottom" title="Delete Data"><i
+            if ($userauth->can('update-companies')) {
+                $button .= ' <a href="' . route('company.edit', ['id' => $data->id]) . '" class="btn btn-sm btn-success action mr-1" data-id=' . $data->id . ' data-type="edit" data-toggle="tooltip" data-placement="bottom" title="Edit Data"><i
+                class="fas fa-pencil-alt"></i></a>';
+            }
+            if ($userauth->can('delete-companies')) {
+                $button .= ' <button  class="btn btn-sm btn-danger  action" data-id=' . $data->id . ' data-type="delete" data-route="' . route('company.delete', ['id' => $data->id]) . '" data-toggle="tooltip" data-placement="bottom" title="Delete Data"><i
                                                             class="fas fa-trash-alt "></i></button>';
+            }
             return '<div class="d-flex gap-2">' . $button . '</div>';
         })->rawColumns(['action'])->make(true);
     }
@@ -64,7 +71,17 @@ class CompanyController extends Controller
             'phone.numeric' => 'Nomor telepon harus berupa angka.',
         ]);
 
-        $company->create($request->all());
+        $company = $company->create($request->all());
+
+        // Log activity for company creation
+        activity()
+            ->causedBy(Auth::user()) // Logs who performed the action
+            ->performedOn($company) // The entity being changed
+            ->event('created') // Event of the action
+            ->withProperties([
+                'attributes' => $company->toArray() // The data that was created
+            ])
+            ->log('Company created');
 
         return redirect()->route('company')->with(['status' => 'Success', 'message' => 'Berhasil Menambahkan Company']);
     }
@@ -82,15 +99,12 @@ class CompanyController extends Controller
         return view('pages.master.company.edit', $data);
     }
 
-
-
     /**
      * Update the specified resource in storage.
      */
     public function update(Request $request, $id)
     {
         try {
-
             $request->validate([
                 'name' => 'required',
                 'address' => 'required',
@@ -103,8 +117,20 @@ class CompanyController extends Controller
             ]);
 
             $company = Company::find($id);
+            $oldCompany = $company->toArray(); // Capture the original data before update
 
             $company->update($request->all());
+
+            // Log activity for company update
+            activity()
+                ->causedBy(Auth::user()) // Logs who performed the action
+                ->performedOn($company) // The entity being changed
+                ->event('updated') // Event of the action
+                ->withProperties([
+                    'old' => $oldCompany, // The data before update
+                    'attributes' => $company->toArray() // The updated data
+                ])
+                ->log('Company updated');
 
             return redirect()->route('company')->with(['status' => 'Success', 'message' => 'Berhasil Mengubah Company']);
         } catch (Exception $e) {
@@ -121,10 +147,21 @@ class CompanyController extends Controller
     public function destroy(string $id)
     {
         try {
-            $data = Company::findOrFail($id);
-            $data->delete();
+            $company = Company::findOrFail($id);
+            $companyData = $company->toArray(); // Capture the data before deletion
+            
+            $company->delete();
 
-            //return response
+            // Log activity for company deletion
+            activity()
+                ->causedBy(Auth::user()) // Logs who performed the action
+                ->performedOn($company) // The entity being changed
+                ->event('deleted') // Event of the action
+                ->withProperties([
+                    'attributes' => $companyData // The data before deletion
+                ])
+                ->log('Company deleted');
+
             return response()->json([
                 'status' => 'success',
                 'success' => true,
