+++ app/Http/Controllers/Master/ItemController.php
@@ -6,6 +6,8 @@ use App\Http\Controllers\Controller;
 use App\Models\Item;
 use App\Models\TypeItem;
 use App\Models\Unit;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
@@ -41,13 +43,15 @@ class ItemController extends Controller
                 return $item->unit ? $item->unit->name : '-';  // Ensure correct access to unit
             })
             ->addColumn('action', function ($data) {
+                $userauth = User::with('roles')->where('id', Auth::id())->first();
                 $button = '';
-                $button .= '<a href="' . route('item.edit', $data->id) . '" class="btn btn-sm btn-success action mr-1" data-id="' . $data->id . '" data-type="edit" data-toggle="tooltip" data-placement="bottom" title="Edit Data">
-                <i class="fas fa-pencil-alt"></i>
-            </a>';
-                $button .= '<button class="btn btn-sm btn-danger action" data-id="' . $data->id . '" data-type="delete" data-route="' . route('item.delete', $data->id) . '" data-toggle="tooltip" data-placement="bottom" title="Delete Data">
-                <i class="fas fa-trash-alt"></i>
-            </button>';
+                if ($userauth->can('update-items')) {
+                    $button .= '<a href="' . route('item.edit', $data->id) . '" class="btn btn-sm btn-success action mr-1" data-id="' . $data->id . '" data-type="edit" data-toggle="tooltip" data-placement="bottom" title="Edit Data"><i class="fas fa-pencil-alt"></i></a>';
+                }
+                if ($userauth->can('delete-items')) {
+                    $button .= '<button class="btn btn-sm btn-danger action" data-id="' . $data->id . '" data-type="delete" data-route="' . route('item.delete', $data->id) . '" data-toggle="tooltip" data-placement="bottom" title="Delete Data">
+                    <i class="fas fa-trash-alt"></i></button>';
+                }
                 return '<div class="d-flex gap-2">' . $button . '</div>';
             })
             ->rawColumns(['action', 'type', 'unit'])
